+++ src/app/api/admin/forms/route.ts
@@ -1,12 +1,15 @@
-import { ZodError } from 'zod'
+import { z, ZodError } from 'zod'
 import { NotFound } from 'payload'
 import { StatusCodes } from 'http-status-codes'
 import { NextRequest, NextResponse } from 'next/server'
-
 import FormService from '@/data-layer/services/FormService'
 import { Security } from '@/business-layer/middleware/Security'
-import { UpdateFormRequestBody } from '@/types/request-models/FormRequest'
 import { UpdateFormData } from '@/types/Collections'
+
+export const UpdateFormRequestBody = z.object({
+  name: z.string().optional(),
+  description: z.string().optional(),
+})
 
 class RouteWrapper {
   /**
