--- src/types/request-models/FormRequest.ts
@@ -1,6 +1,0 @@
-import { z } from 'zod'
-
-export const UpdateFormRequestBody = z.object({
-  name: z.string().optional(),
-  description: z.string().optional(),
-})
