--- src/types/request-models/ProjectRequests.ts
@@ -1,60 +1,0 @@
-import { z } from 'zod'
-import {
-  FormResponseSchema,
-  MediaSchema,
-  UserSchema,
-  ProjectSchema,
-  SemesterSchema,
-} from '../Payload'
-import { ProjectStatus } from '../Project'
-
-export const UpdateProjectRequestBody = z.object({
-  name: z.string().optional(),
-  description: z.string().optional(),
-  deadline: z
-    .string()
-    .datetime({ message: 'Invalid date format, should be in ISO 8601 format' })
-    .optional(),
-  timestamp: z
-    .string()
-    .datetime({ message: 'Invalid date format, should be in ISO 8601 format' })
-    .optional(),
-  clients: z
-    .union([
-      z.array(z.string()).nonempty('At least one client is required'),
-      z.array(UserSchema).nonempty('At least one client is required'),
-    ])
-    .optional(),
-  attachments: z.array(MediaSchema).max(5).optional(),
-})
-
-export const CreateProjectRequestBody = z.object({
-  name: z.string(),
-  description: z.string(),
-  deadline: z
-    .string()
-    .datetime({ message: 'Invalid date format, should be in ISO 8601 format' })
-    .optional(),
-  timestamp: z.string().datetime({ message: 'Invalid date format, should be in ISO 8601 format' }),
-  clients: z.union([
-    z.array(z.string()).nonempty('At least one client is required'),
-    z.array(UserSchema).nonempty('At least one client is required'),
-  ]),
-  attachments: z.array(MediaSchema).max(5).optional(),
-  formResponse: z.union([z.string(), FormResponseSchema]),
-})
-
-export const PatchSemesterProjectRequestBody = z.object({
-  number: z.number().min(1).nullable().optional(),
-  project: z.union([z.string(), ProjectSchema]).optional(),
-  semester: z.union([z.string(), SemesterSchema]).optional(),
-  status: z.nativeEnum(ProjectStatus).optional(),
-  published: z.boolean().optional(),
-})
-
-export const CreateSemesterProjectRequestBody = z.object({
-  number: z.number().min(1).nullable().optional(),
-  project: z.union([z.string(), ProjectSchema]),
-  status: z.nativeEnum(ProjectStatus),
-  published: z.boolean(),
-})
