--- src/types/request-models/UserRequests.ts
@@ -1,10 +1,0 @@
-import { z } from 'zod'
-
-export const UpdateUserRequestBody = z.object({
-  firstName: z.string().optional(),
-  lastName: z.string().optional(),
-  role: z.enum(['admin', 'client', 'student']).optional(),
-  email: z.string().optional(),
-  introduction: z.string().optional(),
-  affiliation: z.string().optional(),
-})
